Refactor the calls to the Github client
There was a bit of duplication in the instantiation of the Github client,
and also a bit of coupling in the `Report` class.

This PR introduces a `RepositoryHelper` with two class methods that simplify
the instantiation of the Github client and the calls to the API, and also
allow to have in one place the configuration for the client (*).

This way `Report` class will always be provided a repository, and takes away
the previous conditional code and the coupling with the Github client.

(*) Currently it feels weird to have this configuration hardcoded, and also
the fact we are using an ENV variable in a class method, but I felt trying
to anticipate anything else might be premature.