Add a configuration loader class
As agreed, I will split into a few smaller PRs the previous big PR with the
proposal for a `checks` architecture.

This is maybe a bit small one but necessaryi, as the checks factory and the
checks collection rely on first having a way to read the checks configuration.

`Configuration` class is just a delegator for a Hashie::Mash object which will
make the loading and accessing to the configuration with method-like syntax
much easier and human-friendly than working with the usual brakets syntax.

Added a `bin/console.rb` script to be able to open a quick debuging console.