Introduce the checks factory
The `ChecksFactory` class is in charge of `registering` checks and
later on `building` them at runtime.

It exposes a simple interface with the following methods:

* `register` will receive two arguments: the type of check to register
(this is meant to be something along the lines `:name_check` or
`:required_file_check`) and their Class.

* `fetch` will receive one argument: the type of check to be retrieved
from the store, returning the check Class if it exists or raising a
`CheckNotFoundError` exception otherwise.

* `build` will receive two or three arguments: the type of check to
build, the metadata for the individual instance of this check (a type
can be instantiated multiple times with different metadata and/or
arguments to suit the different combination of checks) and optionally,
if needed for this check to do their job a map of arguments.
It will internally use `fetch` to lookup the check Class, so will
raise an exception if not found, as described above.

N.B.: `.registered_checks` method might well be private, but I left it
as public to simplify the testing and also because eventually we might
want to have a way to list the registered checks when running the
audit. We can always revisit this decision later on.