Introduce the checks collection
Now that we have a way to read the checks configuration and build them
(with the factory) we have to store the parameterized instance of each of
the checks in some place so we can work with them in bulk and run all of
them against a specific repository.

To do this, we just need an object that behaves like an Array really, and
we accomplish this (in a similar way to the Configuration class) with a
delegator Class, with the initializer creating the array of checks from
the configuration provided.

There is a method `#run` that will iterate through all the checks in the
collection, invoking their `#run` method and passing the `repository`
being audited. The result of each of the checks is added to an array and
returned.

Fixes #12