Create the first basic check (ExampleCheck)
Checks are simple classes inheriting from `BaseCheck` which at least
need to have declared the `register_check` (to register themselves in
the factory) and to have a `#run` method which accepts a `repository`
and runs the expected check against it.

Their initializer (in the superclass) receives the different
configuration for this instance of the check: metadata and (optional)
arguments.

The result of the check run is evaluated and if success, the check
invokes the `#success` method, otherwise it invokes `#failure`
(protected methods inherited from the superclass).

These methods are in charge of building the response (which at the
moment is just a plain hash which will be converted to JSON, but as a
follow up PR this will change to be an Object so we can serialise to
different formats like CSV in the near future).