Remove superseded code and use new architecture
The scripts have been updated to make use of the new config file (for now
the filename is hardcoded, but this will be a command line option soon),
and the new checks collection is passed to the `Report` class.

Removed old code and amendments to the config file to not check for file
presence if we are already checking for content (is redundant) and
added a check for a valid `Dangerfile`.

Fixes #11
Fixes #13