Add a file presence check
This check will trigger a HEAD request against a given filename, as per
configuration, and the result of the `run` method will be based on
the file existing or not in the given repository.

A new helper class has been introduced to encapsulate the operations
with remote files (like this HEAD request).