Extract shared tests from check specs
Some refactor of the check tests as now we are going to have identical
tests for some of the code in each of the checks, so we can make use of
shared examples to tidy up this a little bit.

These shared tests cover the common areas of every check (initialization,
metadata, arguments...) so the individual check specs can focus on the
`#run` method which is different for each check.