Add a file content check
This check will trigger a GET request on a given filename, as per
configuration, and the result of the `run` method will be based on
the file content matching all of the declared regexp matchers in the
configuration file.

As the file content check is an implicit file existence check
(otherwise the body will return the string `404: Not Found`) it is not
really necessary to perform both checks on the same file. Only `content`
check would suffice.