Enable auto_pagination value to be passed
When we check for the last commit in a repo, we
will always want pagination to be turned on, so
that we only see the N most recent commits. So,
we need to be able to control the value of the
auto_pagination flag every time we call the
github_client method.