Add check for presence of at least one file
There are some checks that requires checking for the presence of at least
a file from a given list of filenames.
Specifically this PR ensures the repository is using one of our supported
CI: Travis or Circle.

Moved the `attr_accessor :filename` from `RepoAudit::RepositoryContent`
module into each of the checks that uses it, as it is more clear to have
in the check itself each of the required arguments, and because for this
new check (AtLeastOneFileCheck) we are not using this attribute.

Fixes #16