Update check shared tests to cope with nil args
Until now we only had `checks` with mandatory arguments, so the shared
examples were created assuming this was always the case, but with the
introduction of our first check without arguments (8bdec44) we need to
adapt the shared examples to cope with this.

After a bit of thinking, I realized we don't need explicit tests for
the arguments, as those are needed as part of the check `#run` and thus
are implicitly tested, in order to support this `#run` method.

There is however a minor change to allow any check to be instantiated
only with the metadata, so tests can be less verbose.

Fixes #40