Add class to encapsulate version matching
This class encapsulate operations with versions, making use of the
already available `Gem::Requirement` and `Gem::Version` to create
a version requirement and validate it against a prospective version
that we've extracted from the provided `content` via a regex that
will be configurable in the config file.

We are making use of named regex captures (i.e. `version`) to make
it more explicit and easy to read when we use them in the config file.