Add check for ruby and rails versions
A new kind of Check has been added (`VersionCheck`) to provide us with a
simple framework to check the versioning of different things like in this
case `ruby` and `rails`.

The way it works is very similar to the `FileContentCheck` in which we
declare a list of files to check their content, and regex of the content
to check for, only in this case, these regex allow for capures of a
`version` to be checked against a requirement also declared in the config.

Currently it iterates through the provided list of files, and stops after
it finds the first one that meets the criteria.
If none found, it fails the check.

As it stands right now, it will fail for repositories without Rails (this
repository will fail this check) and also will fail for Python repos for
example, so next step is to create a mechanism to bypass this check if not
needed (i.e. not a Rails repo, so no need to check for Rails version).